+++ zarrs/src/array/chunk_grid/regular.rs
@@ -304,6 +304,7 @@ mod tests {
     use rayon::iter::ParallelIterator;
 
     use super::*;
+    use crate::array::ArrayIndicesTinyVec;
     use crate::array::chunk_grid::ChunkGridTraitsIterators;
     use crate::array_subset::ArraySubset;
 
@@ -435,13 +436,23 @@ mod tests {
         let iter = chunk_grid.iter_chunk_indices();
         assert_eq!(
             iter.collect::<Vec<_>>(),
-            vec![vec![0, 0, 0], vec![0, 0, 1], vec![1, 0, 0], vec![1, 0, 1]]
+            vec![
+                ArrayIndicesTinyVec::Heap(vec![0, 0, 0]),
+                ArrayIndicesTinyVec::Heap(vec![0, 0, 1]),
+                ArrayIndicesTinyVec::Heap(vec![1, 0, 0]),
+                ArrayIndicesTinyVec::Heap(vec![1, 0, 1]),
+            ]
         );
 
         let iter = chunk_grid.par_iter_chunk_indices();
         assert_eq!(
             iter.collect::<Vec<_>>(),
-            vec![vec![0, 0, 0], vec![0, 0, 1], vec![1, 0, 0], vec![1, 0, 1]]
+            vec![
+                ArrayIndicesTinyVec::Heap(vec![0, 0, 0]),
+                ArrayIndicesTinyVec::Heap(vec![0, 0, 1]),
+                ArrayIndicesTinyVec::Heap(vec![1, 0, 0]),
+                ArrayIndicesTinyVec::Heap(vec![1, 0, 1]),
+            ]
         );
 
         let iter = chunk_grid.iter_chunk_subsets();
@@ -460,10 +471,10 @@ mod tests {
         assert_eq!(
             iter.collect::<Vec<_>>(),
             vec![
-                (vec![0, 0, 0], ArraySubset::new_with_ranges(&[0..1, 0..2, 0..3])),
-                (vec![0, 0, 1], ArraySubset::new_with_ranges(&[0..1, 0..2, 3..6])),
-                (vec![1, 0, 0], ArraySubset::new_with_ranges(&[1..2, 0..2, 0..3])),
-                (vec![1, 0, 1], ArraySubset::new_with_ranges(&[1..2, 0..2, 3..6])),
+                (ArrayIndicesTinyVec::Heap(vec![0, 0, 0]), ArraySubset::new_with_ranges(&[0..1, 0..2, 0..3])),
+                (ArrayIndicesTinyVec::Heap(vec![0, 0, 1]), ArraySubset::new_with_ranges(&[0..1, 0..2, 3..6])),
+                (ArrayIndicesTinyVec::Heap(vec![1, 0, 0]), ArraySubset::new_with_ranges(&[1..2, 0..2, 0..3])),
+                (ArrayIndicesTinyVec::Heap(vec![1, 0, 1]), ArraySubset::new_with_ranges(&[1..2, 0..2, 3..6])),
             ]
         );
     }